Fix premature callbacks during question creation
The promise chain in the question creation database accessor method
boke during the answer insertion step. This caused the redirection
to the question administration page to fail if the answer table
did not update in time. The fix in question updated the promise
chain to include the answer insertion step.

This code, along with the rest of the DB code that leverages
promises will need to be revisited and possibly refactored
once I have a better grasp of promises.